requirements.txt
torch
torchvision
timm
tqdm
Pillow
pandas
matplotlib
